requirements.txt
PySide6
Pillow
# watchdog  # 若要做“文件变动自动热重载”，后续再启用
